Remove reference and update unit test in chlid classes
Resolves #21217
